requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.11.18
    # via voyageai
aiolimiter==1.2.1
    # via voyageai
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
attrs==25.3.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
autoevals==0.0.127
    # via life-sciences-rag (pyproject.toml)
beautifulsoup4==4.13.4
    # via life-sciences-rag (pyproject.toml)
braintrust==0.0.201
    # via life-sciences-rag (pyproject.toml)
braintrust-core==0.0.58
    # via
    #   autoevals
    #   braintrust
certifi==2025.4.26
    # via
    #   pinecone
    #   requests
charset-normalizer==3.4.1
    # via requests
chevron==0.14.0
    # via
    #   autoevals
    #   braintrust
exceptiongroup==1.2.2
    # via braintrust
filelock==3.18.0
    # via huggingface-hub
frozenlist==1.6.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.3.2
    # via huggingface-hub
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via braintrust
googleapis-common-protos==1.70.0
    # via
    #   pinecone
    #   protoc-gen-openapiv2
grpcio==1.71.0
    # via pinecone
huggingface-hub==0.30.2
    # via tokenizers
idna==3.10
    # via
    #   requests
    #   yarl
jsonschema==4.23.0
    # via autoevals
jsonschema-specifications==2025.4.1
    # via jsonschema
lz4==4.4.4
    # via pinecone
multidict==6.4.3
    # via
    #   aiohttp
    #   yarl
numpy==2.2.5
    # via voyageai
packaging==25.0
    # via huggingface-hub
pillow==11.2.1
    # via voyageai
pinecone==6.0.2
    # via life-sciences-rag (pyproject.toml)
pinecone-plugin-interface==0.0.7
    # via pinecone
polyleven==0.9.0
    # via autoevals
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
protobuf==5.29.4
    # via
    #   googleapis-common-protos
    #   pinecone
    #   protoc-gen-openapiv2
protoc-gen-openapiv2==0.0.1
    # via pinecone
pydantic==2.11.4
    # via voyageai
pydantic-core==2.33.2
    # via pydantic
python-dateutil==2.9.0.post0
    # via pinecone
python-dotenv==1.1.0
    # via braintrust
python-slugify==8.0.4
    # via braintrust
pyyaml==6.0.2
    # via
    #   autoevals
    #   huggingface-hub
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   braintrust
    #   huggingface-hub
    #   voyageai
rpds-py==0.24.0
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
soupsieve==2.7
    # via beautifulsoup4
sseclient-py==1.8.0
    # via braintrust
tenacity==9.1.2
    # via voyageai
text-unidecode==1.3
    # via python-slugify
tokenizers==0.21.1
    # via voyageai
tqdm==4.67.1
    # via
    #   braintrust
    #   huggingface-hub
typing-extensions==4.13.2
    # via
    #   beautifulsoup4
    #   braintrust
    #   huggingface-hub
    #   pinecone
    #   pydantic
    #   pydantic-core
    #   referencing
    #   typing-inspection
typing-inspection==0.4.0
    # via pydantic
urllib3==2.4.0
    # via
    #   pinecone
    #   requests
voyageai==0.2.4
    # via life-sciences-rag (pyproject.toml)
yarl==1.20.0
    # via aiohttp